fix: style, input, position AuthPage and add Theme
