+++ src/components/Auth/AuthSwitchElement/AuthSwitchElement.style.tsx
@@ -1,9 +1,14 @@
 import { Switch } from '@mui/material';
 import { styled } from '@mui/material/styles';
 
+const numbers = {
+  authHeight: 16,
+  switchThumb: 12,
+};
+
 const AuthSwitch = styled(Switch)(({ theme }) => ({
   width: 28,
-  height: 16,
+  height: numbers.authHeight,
   padding: 0,
   display: 'flex',
   '&:active': {
@@ -21,27 +26,23 @@ const AuthSwitch = styled(Switch)(({ theme }) => ({
       color: '#fff',
       '& + .MuiSwitch-track': {
         opacity: 1,
-        backgroundColor: theme.palette.mode === 'dark' ? '#177ddc' : '#1890ff',
       },
     },
   },
   '& .MuiSwitch-thumb': {
     boxShadow: '0 2px 4px 0 rgb(0 35 11 / 20%)',
-    width: 12,
-    height: 12,
-    borderRadius: 6,
+    width: numbers.switchThumb,
+    height: numbers.switchThumb,
+    borderRadius: numbers.authHeight / 2,
     transition: theme.transitions.create(['width'], {
       duration: 200,
     }),
   },
   '& .MuiSwitch-track': {
-    borderRadius: 16 / 2,
+    borderRadius: numbers.authHeight / 2,
     opacity: 1,
-    backgroundColor:
-      theme.palette.mode === 'dark'
-        ? 'rgba(255,255,255,.35)'
-        : 'rgba(0,0,0,.25)',
     boxSizing: 'border-box',
+    backgroundColor: theme.palette.secondary.main,
   },
 }));
