+++ src/components/Auth/Registration/Registration.tsx
@@ -1,10 +1,11 @@
 import React from 'react';
-import { Box, TextField } from '@mui/material';
+import { Box } from '@mui/material';
 import { useTranslation } from 'react-i18next';
 import { useAppDispatch } from '../../../store/hooks';
 import { signin } from '../../../store/auth/authSlice';
 import RoutesList from '../../../navigation/routes';
-import { SubmitButton, SubmitLink } from '../Auth.style';
+import Input from '../../../common/Input/Input';
+import { SubmitLink, SubmitButton } from '../Auth.style';
 
 function Registration(): JSX.Element {
   const [t] = useTranslation();
@@ -16,40 +17,26 @@ function Registration(): JSX.Element {
 
   return (
     <Box component="form" onSubmit={handleSubmit} noValidate>
-      <TextField
-        margin="normal"
-        required
-        fullWidth
+      <Input
         id="email"
         label={t('auth.email')}
         name="email"
         autoComplete="email"
-        autoFocus
       />
-      <TextField
-        margin="normal"
-        required
-        fullWidth
+      <Input
         id="reemail"
         label={t('auth.reemail')}
         name="reemail"
         autoComplete="reemail"
-        autoFocus
       />
-      <TextField
-        margin="normal"
-        required
-        fullWidth
+      <Input
         name="password"
         label={t('auth.password')}
         type="password"
         id="password"
         autoComplete="current-password"
       />
-      <TextField
-        margin="normal"
-        required
-        fullWidth
+      <Input
         name="repassword"
         label={t('auth.repassword')}
         type="password"
